chore(types): move some common types to main.d.ts
